// src/types/sales.ts
export interface SalesDataItem {
  id: string;
  productName?: string;
  sku?: string;
  salesLocation?: string;
  price: number;
  demand?: string;
  season?: string;
  month?: string;
  date?: string;
  quantity: number;
  category: string;
  margin: number;
  trend?: string;
  imageUrl?: string;
}